+++ lib/user_screen.dart
@@ -4,7 +4,8 @@ import 'package:fbase/cards.dart';
 import 'package:fbase/logging_in/user_logging_in.dart';
 import 'package:fbase/qrscanner.dart';
 import 'package:fbase/basket.dart';
-import 'package:fbase/user_page.dart';
+import 'package:fbase/selectCafe.dart';
+import 'package:fbase/user_table_page.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter_animated_icons/icons8.dart';
@@ -18,38 +19,6 @@ void updateQR(String newQR) {
   qrCode = newQR;
 }
 
-Future<void> empty(String qrtext) async {
-  // Firebase Firestore bağlantısını al
-  final FirebaseFirestore firestore = FirebaseFirestore.instance;
-
-  // Belge referansını al
-  final DocumentReference docRef = firestore.collection('Masalar').doc(qrtext);
-
-  try {
-    // Belgeyi getir
-    final DocumentSnapshot document = await docRef.get();
-
-    if (document.exists) {
-      // "chairs" alanını güncelle
-      final List<dynamic> chairs =
-          List.from(document.get('chairStatusList') as List<dynamic>);
-
-      // False olan bir sandalye bul
-      int indexToUpdate = chairs.indexWhere((chair) => chair == true);
-
-      if (indexToUpdate != -1) {
-        // İlgili sandalyeyi false yap
-        chairs[indexToUpdate] = false;
-
-        // Güncellenmiş sandalye listesini Firestore'a kaydet
-        await docRef.update({'chairStatusList': chairs});
-      }
-    }
-  } catch (e) {
-    // Hata durumunda ilgili işlemleri gerçekleştir
-    print('Hata: $e');
-  }
-}
 
 class Kullanici extends StatefulWidget {
   final email;
@@ -94,7 +63,8 @@ class _KullaniciState extends State<Kullanici> {
       body: IndexedStack(
         index: selectedIndex,
         children: [
-          masa(docname: docname),
+          //masa(docname: docname),
+          SelectCafe(),
           cuzdan(email: widget.email, docname: docname),
           Cafe(docname: docname),
           hesap(email: widget.email),
@@ -143,196 +113,13 @@ class _KullaniciState extends State<Kullanici> {
   }
 }
 
+
 class chosentable extends StatelessWidget {
   const chosentable({super.key});
 
   @override
   Widget build(BuildContext context) {
     return const Placeholder();
-  }
-}
-
-class masa extends StatefulWidget {
-  final docname;
-
-  masa({Key? mykey, this.docname}) : super(key: mykey);
-
-  @override
-  State<masa> createState() => _masaState();
-}
-
-class _masaState extends State<masa> {
-  TextEditingController yorum = TextEditingController();
-  List<bool> starColors = List.filled(5, false);
-
-
-
-  func() {}
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            const Expanded(
-              child: UserPage(),
-            ),
-            const Divider(thickness: 2),
-            ElevatedButton(
-              onPressed: () {
-                showDialog(
-                  context: context,
-                  builder: (BuildContext context) {
-                    return AlertDialog(
-                      title: Text("Empty Table"),
-                      content: Column(
-                        mainAxisSize: MainAxisSize.min,
-                        children: selectedTables.entries.map((entry) {
-                          int tableIndex = entry.key;
-                          String tableName = entry.value;
-                          return ListTile(
-                            title: Text('$tableName'),
-                            trailing: IconButton(
-                              icon: Icon(Icons.delete),
-                              onPressed: () {
-                                showDialog(
-                                  context: context,
-                                  builder: (BuildContext context) {
-                                    return AlertDialog(
-                                      title:
-                                      Text("Are you sure you want to empty the table?"),
-                                      actions: [
-                                        ElevatedButton(
-                                          onPressed: () {
-                                            setState(() {
-                                              selectedTables.remove(tableIndex);
-                                            });
-                                            empty(tableName);
-                                            Navigator.pop(context);
-                                            Navigator.pop(context);
-                                          },
-                                          child: Text("Yes"),
-                                        ),
-                                        ElevatedButton(
-                                          onPressed: () {
-                                            Navigator.pop(context);
-                                          },
-                                          child: Text("No"),
-                                        ),
-                                      ],
-                                    );
-                                  },
-                                );
-                              },
-                            ),
-                          );
-                        }).toList(),
-                      ),
-                      actions: [
-                        ElevatedButton(
-                          onPressed: () {
-                            Navigator.pop(context);
-                          },
-                          child: Text("Cancel"),
-                        ),
-                      ],
-                    );
-                  },
-                );
-              },
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.lightBlueAccent,
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(32.0),
-                ),
-              ),
-              child: const Row(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  Icon(
-                    Icons.delete,
-                  ),
-                  SizedBox(width: 5),
-                  Text("Empty table"),
-                ],
-              ),
-            ),
-            ElevatedButton.icon(
-              onPressed: () {
-                Navigator.of(context).push(MaterialPageRoute(builder: (context) => QRScanner()));
-              },
-              icon: Icon(
-                Icons.camera_alt,
-                color: Colors.white,
-              ),
-              label: Text('Get a table'),
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.deepOrange,
-                foregroundColor: Colors.white,
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(32.0),
-                ),
-              ),
-            ),
-            ElevatedButton.icon(
-              onPressed: () {
-                Navigator.of(context).push(MaterialPageRoute(builder: (context) => yorumSor()));
-              },
-              icon: Icon(
-                Icons.comment,
-                color: Colors.white,
-              ),
-              label: Text('Comment'),
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Colors.green,
-                foregroundColor: Colors.white,
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(32.0),
-                ),
-              ),
-            ),
-
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget yorumSor() {
-    TextEditingController yorum = TextEditingController();
-    List<bool> starColors = List.generate(5, (index) => false);
-
-    return Scaffold(
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          TextField(
-            controller: yorum,
-          ),
-          RatingBar.builder(
-            minRating: 1,
-            itemSize: 46,
-            itemPadding: EdgeInsets.symmetric(horizontal: 4),
-            itemBuilder: (context, index) => Icon(
-              Icons.star,
-              color: starColors[index] ? Colors.amber : Colors.grey,
-            ),
-            updateOnDrag: true,
-            onRatingUpdate: (rating) {
-              setState(() {
-                starColors = List.generate(5, (index) => index < rating.round());
-                FirebaseFirestore.instance
-                    .collection('Yildizlar')
-                    .doc(widget.docname)
-                    .set({'yildizlar': starColors, 'yorum': yorum.text});
-              });
-            },
-          ),
-        ],
-      ),
-    );
   }
 }
